Moved component test to unit test
