--- src/engine/source/policy/test/src/unit/policyRep_test.cpp
@@ -1,5 +1,0 @@
-#include <gtest/gtest.h>
-
-TEST(PolicyRepUTest, uTest) {
-  GTEST_SKIP();
-}
